19.09.21

+++ app/src/main/java/de/luki2811/dev/vokabeltrainer/MainActivity.java
@@ -11,10 +11,12 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.cardview.widget.CardView;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.content.ContextCompat;
+
+import com.google.android.material.button.MaterialButton;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -23,6 +25,8 @@ import org.json.JSONObject;
 import java.io.File;
 
 public class MainActivity extends AppCompatActivity {
+
+    final static String LEKTION_NAME = "de.luki2811.dev.vokabeltrainer";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,7 +58,7 @@ public class MainActivity extends AppCompatActivity {
                     cardView.setLayoutParams(layoutparams);
                     cardView.setRadius(25);
                     cardView.setContentPadding(10,10,10,10);
-                    cardView.setCardBackgroundColor(getColor(R.color.Black));
+                    cardView.setCardBackgroundColor(getColor(R.color.Aquamarine));
                     cardView.setCardElevation(3);
                     cardView.setMaxCardElevation(5);
 
@@ -76,48 +80,71 @@ public class MainActivity extends AppCompatActivity {
                     textInCard.setTextColor(Color.WHITE);
                     textInCard.setGravity(Gravity.TOP);
 
+                    // H+W for icons
+                    RelativeLayout.LayoutParams layoutparamsIcons = new RelativeLayout.LayoutParams(100,100);
 
+                    // Delete button
                     ImageButton delete = new ImageButton(this);
                     delete.setBackgroundResource(R.drawable.rounded_red_button);
                     delete.setImageDrawable(ContextCompat.getDrawable(getApplicationContext(),R.drawable.outline_delete_24));
-                    RelativeLayout.LayoutParams layoutparamsDelete = new RelativeLayout.LayoutParams(100,100);
-                    delete.setLayoutParams(layoutparamsDelete);
+                    delete.setLayoutParams(layoutparamsIcons);
 
-                    delete.setOnClickListener(view -> {
-                        File file = new File(getApplicationContext().getFilesDir(), textInCard.getText() + ".json");
-                        if(file.exists()){
-                            boolean deleted = file.delete();
-                            if(deleted){
-                                try {
-                                    for(int i2 = 0; i2 <= indexArrayJson.length() - 1; i2++){
-                                        if(indexArrayJson.getJSONObject(i2).getString("name").contentEquals(textInCard.getText())){
-                                            System.out.println(indexArrayJson.toString());
-                                            indexArrayJson.remove(i2);
-
-                                            indexJson.put("index", indexArrayJson);
-                                            indexDatei.writeInFile(indexJson.toString(),this);
-
-                                            cardView.setVisibility(View.INVISIBLE);
-                                            cardView.setLayoutParams(new LinearLayout.LayoutParams(0,0));
-                                            Toast.makeText(getApplicationContext(), getString(R.string.deleted_succesfull), Toast.LENGTH_SHORT).show();
+                    delete.setOnClickListener(view -> new AlertDialog.Builder(this)
+                            .setTitle("")
+                            .setMessage("Möchtest du wirklich die Lektion löschen")
+                            .setIcon(R.drawable.outline_delete_24)
+                            .setPositiveButton(R.string.delete, (dialogInterface, i1) -> {
+                                File file = new File(getApplicationContext().getFilesDir(), textInCard.getText() + ".json");
+                                if(file.exists()){
+                                    boolean deleted = file.delete();
+                                    if(deleted){
+                                        try {
+                                            for(int i2 = 0; i2 <= indexArrayJson.length() - 1; i2++){
+                                                if(indexArrayJson.getJSONObject(i2).getString("name").contentEquals(textInCard.getText())){
+                                                    System.out.println(indexArrayJson.toString());
+                                                    indexArrayJson.remove(i2);
+                                                    indexJson.put("index", indexArrayJson);
+                                                    indexDatei.writeInFile(indexJson.toString(),getApplicationContext());
+                                                    cardView.setVisibility(View.INVISIBLE);
+                                                    cardView.setLayoutParams(new LinearLayout.LayoutParams(0,0));
+                                                    Toast.makeText(getApplicationContext(), getString(R.string.deleted_succesfull), Toast.LENGTH_SHORT).show();
+                                                }
+                                            }
+                                        } catch (JSONException e){
+                                            e.printStackTrace();
                                         }
                                     }
-
-                                } catch (JSONException e){
-                                    e.printStackTrace();
                                 }
-                            }
-
-                        }
-
+                            })
+                            .setNegativeButton(R.string.cancel, (dialogInterface, i12) -> Toast.makeText(getApplicationContext(), getString(R.string.cancel), Toast.LENGTH_SHORT).show())
+                            .show());
+                    // Edit Button
+                    ImageButton cardEdit = new ImageButton(this);
+                    cardEdit.setBackgroundResource(R.drawable.rounded_blue_button);
+                    cardEdit.setImageDrawable(ContextCompat.getDrawable(getApplicationContext(),R.drawable.ic_outline_edit_24));
+                    cardEdit.setLayoutParams(layoutparamsIcons);
+                    cardEdit.setOnClickListener(view -> {
+                        Intent intent = new Intent(this, EditLektionActivity.class);
+                        intent.putExtra(LEKTION_NAME, textInCard.getText());
+                        startActivity(intent);
+                    });
+                    // Train Button
+                    MaterialButton cardLearnButton = new MaterialButton(this, null, R.attr.borderlessButtonStyle);
+                    cardLearnButton.setText(R.string.practice);
+                    cardLearnButton.setBackgroundDrawable(getDrawable(R.drawable.outline_button));
+                    cardLearnButton.setCornerRadius(100);
+                    cardLearnButton.setOnClickListener(view -> {
 
                     });
 
                     // Add all to a Layout
+                    // TEMP without structure
                     LinearLayout cardLayout = new LinearLayout(this);
 
                     cardLayout.addView(delete);
+                    cardLayout.addView(cardEdit);
                     cardLayout.addView(textInCard);
+                    cardLayout.addView(cardLearnButton);
 
                     cardView.addView(cardLayout);
 
@@ -127,12 +154,6 @@ public class MainActivity extends AppCompatActivity {
                 e.printStackTrace();
             }
         }
-
-
-
-
-
-
     }
 
     public void createNewLektion(View view){
